# docs/swagger.yaml
basePath: /
definitions:
  dto.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  dto.DetailPostResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/dto.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      like_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  dto.LikeOrUnlikeCommentRequest:
    properties:
      comment_id:
        type: integer
    required:
    - comment_id
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RefreshTokenResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: hi@devbercerita.local
    name: DevBercerita Team
  description: 'API untuk platform DevBercerita: register/login, posting, dan komentar.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: DevBercerita API
  version: "1.0"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange an existing refresh token (via Authorization header) for
        new tokens.
      parameters:
      - description: Refresh token payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Refresh session tokens
      tags:
      - Auth
...
  /comments/action:
    post:
      consumes:
      - application/json
      parameters:
      - description: Comment like payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LikeOrUnlikeCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      security:
      - BearerAuth: []
      summary: Like or unlike a comment
      tags:
      - Comments
...
  /posts/{post_id}/detail:
    get:
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DetailPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MessageResponse'
      summary: Get post detail
      tags:
      - Posts
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"